# openapi.yaml
openapi: 3.0.3
info:
  title: MVP Launch API
  description: |
    API for MVP Launch platform - data ingestion, forecasting, and visualization.
  version: 1.0.0
  contact:
    name: Forge Intelligence
    email: dev@forgeintelligence.com
servers:
  - url: https://api.forgeintelligence.com/api/v1
    description: Production server
  - url: http://localhost:8000/api/v1
    description: Local development server
security:
  - bearerAuth: []
...
paths:
...
  /nl-query/translate:
    post:
      tags:
        - NL Query
      summary: Translate natural language to SQL and execute
      operationId: translateNLQuery
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
                - data_source_id
                - user_id
              properties:
                query:
                  type: string
                  description: Natural language question
                data_source_id:
                  type: string
                  description: ID of the connected data source
                user_id:
                  type: string
                  description: ID of the user making the query
                execute:
                  type: boolean
                  default: true
                  description: Whether to execute the generated SQL
                max_rows:
                  type: integer
                  default: 10000
                  description: Maximum rows to return
      responses:
        '200':
          description: Query translated and optionally executed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  natural_language_query:
                    type: string
                  generated_sql:
                    type: string
                  executed_sql:
                    type: string
                  results:
                    type: array
                    items:
                      type: object
                  row_count:
                    type: integer
                  confidence_score:
                    type: integer
                  confidence_level:
                    type: string
                    enum: [high, medium, low]
                  confidence_indicator:
                    type: string
                  follow_up_questions:
                    type: array
                    items:
                      type: string
                  error_message:
                    type: string
                  execution_time_ms:
                    type: integer
                  status:
                    type: string
        '404':
          description: Data source not found
        '500':
          description: Translation or execution failed
  /nl-query/suggest-rephrase:
    post:
      tags:
        - NL Query
      summary: Suggest rephrase for failed query
      operationId: suggestRephrase
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - query
              - error_message
              properties:
                query:
                  type: string
                error_message:
                  type: string
      responses:
        '200':
          description: Rephrase suggestions
          content:
            application/json:
              schema:
                type: object
                properties:
                  original_query:
                    type: string
                  suggestions:
                    type: array
                    items:
                      type: string
                  reason:
                    type: string
  /nl-query/schema:
    get:
      tags:
        - NL Query
      summary: Get database schema for a data source
      operationId: getNLQuerySchema
      security:
        - bearerAuth: []
      parameters:
        - name: data_source_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Schema information
        '404':
          description: Data source not found
  /nl-query/history:
    get:
      tags:
        - NL Query
      summary: Get NL query history for a user
      operationId: getNLQueryHistory
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: data_source_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Query history
  /nl-query/history/{queryId}:
    get:
      tags:
        - NL Query
      summary: Get a specific query from history
      operationId: getNLQueryById
      security:
        - bearerAuth: []
      parameters:
        - name: queryId
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Query details
        '404':
          description: Query not found
    delete:
      tags:
        - NL Query
      summary: Delete a query from history
      operationId: deleteNLQuery
      security:
        - bearerAuth: []
      parameters:
        - name: queryId
          in: path
          required: true
          schema:
            type: string
        - name: user_id
          in: query
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Query deleted
        '404':
          description: Query not found
  /nl-query/recent:
    get:
      tags:
        - NL Query
      summary: Get recent queries for sidebar
      operationId: getRecentNLQueries
      security:
        - bearerAuth: []
      parameters:
        - name: user_id
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Recent queries
  /nl-query/confidence:
    post:
      tags:
        - NL Query
      summary: Get confidence score without executing
      operationId: getNLQueryConfidence
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
                - data_source_id
              properties:
                query:
                  type: string
                data_source_id:
                  type: string
                user_id:
                  type: string
      responses:
        '200':
          description: Confidence score
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...